+++ changelog.md
@@ -1,2 +1,4 @@
+- <a href="https://github.com/groupby/issues/issues/720">iss2</a> Quickstart to support more refinements (and PHP API)
+
 Changelog
 ---